Moving String Extensions to `Micorosoft.Toolkit` as suggesred
